SimpleRoots.jl

SimpleRoots.Both — Type

Return both lower and upper result values as a tuple

SimpleRoots.Lower — Type

Return the lower result value

SimpleRoots.QuadraticResult — Type

Used to specifiy the result of quad, either Lower, Upper or Both

SimpleRoots.Secant — Type

Secant(bracket)
Secant(bracket...)

Secant method of bracketed root-finding.

Bracket can be a length 2 vector, a tuple or 2 argument

SimpleRoots.Upper — Type

Return the upper result value

SimpleRoots.findzero — Method

findzero(f, bracket; atol=1e-7, max_iter=100)
findzero(f, method::AbstractBisection; atol=1e-7, max_iter=100)

Find roots using the passed in method. If a bracket is passed in without specifying the method, use the Secant method.

SimpleRoots.findzero — Method

findzero(f, method::Bisection, atol, max_iter)
findzero(f, method::Bisection; atol=1e-7, max_iter=100)

Find root using the bisection method.

SimpleRoots.findzero — Method

findzero(f, method::Brent, atol, maxiter)
findzero(f, method::Brent; atol=1e-7, max_iter=100)

Find root using Brents method. Returns a tuple of the found value and a Bool specifying wether the root was found within the tolerance, or not.

SimpleRoots.findzero — Method

findzero(f, method::Secant, atol, max_iter)
findzero(f, method::Secant; atol=1e-7, max_iter=100)

Find root using the secant method.

SimpleRoots.quad — Method

function quad(res::QuadraticResult, a, b, c)

Simple quadratic equation solver.

Returns the result specified with res Lower, Upper or Both